backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
alembic
slowapi
bcrypt
groq
pydantic
reportlab
sentence-transformers
numpy
